+++ src/App.tsx
@@ -9,6 +9,7 @@ import {
 import "./App.css"
 import { Navbar } from "@/components/Navbar"
 import { useAppFontSize } from "@/features/DisplaySettings/fontSize"
+import { useInputCapitalization } from "@/features/DisplaySettings/inputCapitalization"
 import { HomeView } from "@/views/HomeView"
 import { FirmSettingsView } from "@/views/FirmSettingsView"
 import { ConfigurationView } from "@/views/ConfigurationView"
@@ -19,6 +20,7 @@ import { RequestDetailsView } from "@/views/RequestDetailsView"
 
 function AppLayout() {
   useAppFontSize()
+  useInputCapitalization()
   const location = useLocation()
   const isRepairsList = location.pathname === "/naprawy"
   const isRepairCard =